bdgraph.npn {BDgraph}	R Documentation

Nonparametric transfer

Description

Transfers non-Gaussian data to Gaussian.

Usage

 bdgraph.npn( data, npn = "shrinkage", npn.thresh = NULL ) 

Arguments

data
An (n x p) matrix or a data.frame corresponding to the data (n is the sample size and p is the number of variables).

npn
A character with three options "shrinkage" (default), "truncation", and "skeptic". Option "shrinkage" is for the shrunken transformation, option "truncation" is for the truncated transformation and option "skeptic" is for the non-paranormal skeptic transformation. For more details see references.

npn.thresh
The truncation threshold; it is only for the truncated transformation (npn= "truncation"). The default value is 1/(4n^{1/4} √{π \log(n)}).

Value

data
An (n \times p) matrix of transferred data, if npn = "shrinkage" or "truncation", and a non-paranormal correlation (p \times p) matrix, if npn = "skeptic".

Author(s)

Reza Mohammadi

References

Liu, H., F. Han, M. Yuan, J. Lafferty, and L. Wasserman (2012). High Dimensional Semiparametric Gaussian Copula Graphical Models, Annals of Statistics 40(4):2293-2326

Zhao, T. and H. Liu (2012). The huge Package for High-dimensional Undirected Graph Estimation in R, Journal of Machine Learning Research, 13:1059-1062

See Also

bdgraph.sim and bdgraph

Examples

## Not run: 
# Generating multivariate normal data from a 'random' graph
data.sim <- bdgraph.sim( n = 6, p = 4, size = 4 )
data     <- ( data.sim $ data - 3 ) ^ 4
data
   
# Transfer the data by truncation 
bdgraph.npn( data, npn = "truncation" )
  
# Transfer the data by shrunken 
bdgraph.npn( data, npn = "shrunken" )
  
# Transfer the data by skeptic 
bdgraph.npn( data, npn = "skeptic" )

## End(Not run)

[Package BDgraph version 2.49 Index]