churn {BDgraph}	R Documentation

Churn data set

Description

The data can be downloaded from IBM Sample Data Sets. Customer churn occurs when customers stop doing business with a company, also known as customer attrition. The data set contains 3333 rows (customers) and 20 columns (features). The "Churn" column is our target which indicate whether customer churned (left the company) or not.

Usage

 data( churn ) 

Format

The churn dataset, as a data frame, contains 3333 rows (customers) and 20 columns (variables/features). The 20 variables are:

State: Categorical, for the 50 states and the District of Columbia.

Account.Length: count, how long account has been active.

Area.Code: Categorical.

Int.l.Plan: Categorical, yes or no, international plan.

VMail.Plan: Categorical, yes or no, voice mail plan.

VMail.Message: Count, number of voice mail messages.

Day.Mins: Continuous, minutes customer used service during the day.

Day.Calls: Count, total number of calls during the day.

Day.Charge: Continuous, total charge during the day.

Eve.Mins: Continuous, minutes customer used service during the evening.

Eve.Calls: Count, total number of calls during the evening.

Eve.Charge: Continuous, total charge during the evening.

Night.Mins: Continuous, minutes customer used service during the night.

Night.Calls: Count, total number of calls during the night.

Night.Charge: Continuous, total charge during the night.

Intl.Mins: Continuous, minutes customer used service to make international calls.

Intl.Calls: Count, total number of international calls.

Intl.Charge: Continuous, total international charge.

CustServ.Calls: Count, number of calls to customer service.

Churn: Categorical, True or False. Indicator of whether the customer has left the company (True or False).

References

Larose, D. T. and Larose, C. D. (2014). Discovering knowledge in data: an introduction to data mining. John Wiley & Sons.

Examples

data( churn )

summary( churn )

[Package BDgraph version 2.62 Index]