compare {BDgraph}	R Documentation

Graph structure comparison

Description

This function provides several measures to assess the performance of the graphical structure learning.

Usage

 compare( sim.obj, bdgraph.obj, bdgraph.obj2 = NULL, bdgraph.obj3 = NULL, 
          colnames = NULL, vis = FALSE ) 

Arguments

sim.obj
An object with S3 class "sim" from function bdgraph.sim. It also can be the adjacency matrix corresponding to the true graph structure.

bdgraph.obj
An object with S3 class "bdgraph" from function bdgraph. It also can be an adjacency matrix corresponding to an estimated graph.

bdgraph.obj2
An object with S3 class "bdgraph" from function bdgraph. It also can be an adjacency matrix corresponding to an estimated graph. It is for comparing two different approaches.

bdgraph.obj3
An object with S3 class "bdgraph" from function bdgraph. It also can be an adjacency matrix corresponding to an estimated graph. It is for comparing three different approaches.

colnames
A character vector giving the column names for the result table.

vis
Visualize the true graph and estimated graph structures.

Value

True positive
The number of correctly estimated links.

True negative
The number of true non-existing links which is correctly estimated.

False positive
The number of links which they are not in the true graph, but are incorrectly estimated.

False negative
The number of links which they are in the true graph, but are not estimated.

F1-score
A weighted average of the "positive predictive" and "true positive rate". The F1-score value reaches its best value at 1 and worst score at 0.

Specificity
The Specificity value reaches its best value at 1 and worst score at 0.

Sensitivity
The Sensitivity value reaches its best value at 1 and worst score at 0.

MCC
The Matthews Correlation Coefficients (MCC) value reaches its best value at 1 and worst score at 0.

Author(s)

Reza Mohammadi, Antonio Abbruzzo, and Ivan Vujacic

References

Mohammadi, A. and E. Wit (2015). Bayesian Structure Learning in Sparse Gaussian Graphical Models, Bayesian Analysis, 10(1):109-138

Mohammadi, A. and E. Wit (2015). BDgraph: An R Package for Bayesian Structure Learning in Graphical Models, arXiv preprint arXiv:1501.05108

Mohammadi, A. et al (2017). Bayesian modelling of Dupuytren disease by using Gaussian copula graphical models, Journal of the Royal Statistical Society: Series C

Mohammadi, A., Massam H., and G. Letac (2017). The Ratio of Normalizing Constants for Bayesian Graphical Gaussian Model Selection, arXiv preprint arXiv:1706.04416

Dobra, A. and A. Mohammadi (2017). Loglinear Model Selection and Human Mobility, arXiv preprint arXiv:1711.02623

See Also

bdgraph and select

Examples

## Not run: 
# Generating multivariate normal data from a 'random' graph
data.sim <- bdgraph.sim( n = 50, p = 6, size = 7, vis = TRUE )
    
# Running sampling algorithm based on GGMs 
sample.ggm <- bdgraph( data = data.sim, method = "ggm", iter = 10000 )
# Comparing the results
compare( data.sim, sample.ggm, colnames = c( "True", "GGM" ), vis = TRUE )
    
# Running sampling algorithm based on GCGMs
sample.gcgm <- bdgraph( data = data.sim, method = "gcgm", iter = 10000 )
# Comparing GGM and GCGM methods
compare( data.sim, sample.ggm, sample.gcgm, colnames = c("True", "GGM", "GCGM"), vis = TRUE )

## End(Not run)

[Package BDgraph version 2.48 Index]