gnorm {BDgraph}	R Documentation

Normalizing constant for G-Wishart

Description

Calculates log of the normalizing constant of G-Wishart distribution based on the Monte Carlo method, developed by Atay-Kayis and Massam (2005).

Usage

gnorm( adj.g, b = 3, D = diag( ncol( adj.g ) ), iter = 100 )

Arguments

adj.g
The adjacency matrix corresponding to the graph structure. It is an upper triangular matrix in which aij = 1 if there is a link between notes i and j, otherwise aij = 0.

b
The degree of freedom for G-Wishart distribution, W_G(b, D).

D
The positive definite (p \times p) "scale" matrix for G-Wishart distribution, W_G(b,D). The default is an identity matrix.

iter
The number of iteration for the Monte Carlo approximation.

Details

Log of the normalizing constant approximation using Monte Carlo method for a G-Wishart distribution, K \sim W_G(b, D), with density:

Pr(K) = \frac{1}{I(b, D)} |K| ^ {(b - 2) / 2} \exp ≤ft\{- \frac{1}{2} \mbox{trace}(K \times D)\right\}.

Value

Log of the normalizing constant of G-Wishart distribution.

Author(s)

Reza Mohammadi and Ernst Wit

References

Atay-Kayis, A. and H. Massam (2005). A monte carlo method for computing the marginal likelihood in nondecomposable Gaussian graphical models, Biometrika, 92(2):317-335

Mohammadi, A. and E. Wit (2015). Bayesian Structure Learning in Sparse Gaussian Graphical Models, Bayesian Analysis, 10(1):109-138

Mohammadi, A., Massam H., and G. Letac (2017). The Ratio of Normalizing Constants for Bayesian Graphical Gaussian Model Selection, arXiv preprint arXiv:1706.04416

Examples

# adj.g: adjacency matrix of graph with 3 nodes and 2 links
adj.g <- matrix( c( 0, 0, 1,
                    0, 0, 1,
                    0, 0, 0 ), 3, 3, byrow = TRUE )		                
   
gnorm( adj.g, b = 3, D = diag( 3 ) )

[Package BDgraph version 2.48 Index]