pgraph {BDgraph}	R Documentation

Posterior probabilities of the graphs

Description

Provides the estimated posterior probabilities for the most likely graphs or a specific graph.

Usage

 pgraph( bdgraph.obj, number.g = 4, adj = NULL ) 

Arguments

bdgraph.obj
An object of S3 class "bdgraph", from function bdgraph.

number.g
The number of graphs with the highest posterior probabilities to be shown. This option is ignored if 'adj' is specified.

adj
An adjacency matrix corresponding to a graph structure. It is an upper triangular matrix in which a_{ij}=1 if there is a link between notes i and j, otherwise a_{ij}=0. It also can be an object of S3 class "sim", from function bdgraph.sim.

Value

selected_g
the adjacency matrices which corresponding to the graphs with the highest posterior probabilities.

prob_g
A vector of the posterior probabilities of the graphs corresponding to 'selected\_g'.

Author(s)

Reza Mohammadi a.mohammadi@uva.nl and Ernst Wit

References

Mohammadi, R. and Wit, E. C. (2019). BDgraph: An R Package for Bayesian Structure Learning in Graphical Models, Journal of Statistical Software, 89(3):1-30

Mohammadi, A. and Wit, E. C. (2015). Bayesian Structure Learning in Sparse Gaussian Graphical Models, Bayesian Analysis, 10(1):109-138

Dobra, A. and Mohammadi, R. (2018). Loglinear Model Selection and Human Mobility, Annals of Applied Statistics, 12(2):815-845

Letac, G., Massam, H. and Mohammadi, R. (2018). The Ratio of Normalizing Constants for Bayesian Graphical Gaussian Model Selection, arXiv preprint arXiv:1706.04416v2

Mohammadi, A. et al (2017). Bayesian modelling of Dupuytren disease by using Gaussian copula graphical models, Journal of the Royal Statistical Society: Series C, 66(3):629-645

See Also

bdgraph, bdgraph.mpl

Examples

## Not run: 
# Generating multivariate normal data from a 'random' graph
data.sim <- bdgraph.sim( n = 50, p = 6, size = 6, vis = TRUE )
   
bdgraph.obj <- bdgraph( data = data.sim, save = TRUE )
   
# Estimated posterior probability of the true graph
pgraph( bdgraph.obj, adj = data.sim )
   
# Estimated posterior probability of first and second graphs with highest probabilities
pgraph( bdgraph.obj, number.g = 2 )

## End(Not run)

[Package BDgraph version 2.62 Index]