plinks {BDgraph}	R Documentation

Estimated posterior link probabilities

Description

Provides the estimated posterior link probabilities for all possible links in the graph.

Usage

 plinks( bdgraph.obj, round = 2, burnin = NULL ) 

Arguments

bdgraph.obj
An object of S3 class "bdgraph", from function bdgraph. It also can be an object of S3 class "ssgraph", from the function ssgraph of R package ssgraph.

round
A value for rounding all probabilities to the specified number of decimal places.

burnin
The number of burn-in iteration to scape.

Value

An upper triangular matrix which corresponds the estimated posterior probabilities for all possible links.

Author(s)

Reza Mohammadi a.mohammadi@uva.nl and Ernst Wit

References

Mohammadi, R. and Wit, E. C. (2019). BDgraph: An R Package for Bayesian Structure Learning in Graphical Models, Journal of Statistical Software, 89(3):1-30

Mohammadi, A. and Wit, E. C. (2015). Bayesian Structure Learning in Sparse Gaussian Graphical Models, Bayesian Analysis, 10(1):109-138

Letac, G., Massam, H. and Mohammadi, R. (2018). The Ratio of Normalizing Constants for Bayesian Graphical Gaussian Model Selection, arXiv preprint arXiv:1706.04416v2

Dobra, A. and Mohammadi, R. (2018). Loglinear Model Selection and Human Mobility, Annals of Applied Statistics, 12(2):815-845

Mohammadi, A. et al (2017). Bayesian modelling of Dupuytren disease by using Gaussian copula graphical models, Journal of the Royal Statistical Society: Series C, 66(3):629-645

See Also

bdgraph, bdgraph.mpl

Examples

## Not run: 
# Generating multivariate normal data from a 'circle' graph
data.sim <- bdgraph.sim( n = 70, p = 6, graph = "circle", vis = TRUE )

bdgraph.obj   <- bdgraph( data = data.sim, iter = 10000 )

plinks( bdgraph.obj, round = 2 )

## End(Not run)

[Package BDgraph version 2.62 Index]