plot.bdgraph {BDgraph}	R Documentation

Plot function for S3 class "bdgraph"

Description

Visualizes structure of the selected graphs which could be a graph with links for which their estimated posterior probabilities are greater than 0.5 or graph with the highest posterior probability.

Usage

 ## S3 method for class 'bdgraph'
plot( x, cut = NULL, number.g = 1, layout = layout.circle, ... ) 

Arguments

x
An object of S3 class "bdgraph", from function bdgraph.

cut
Threshold for including the links in the selected graph based on the estimated posterior probabilities of the links; See the examples.

number.g
The number of graphs with the highest probabilities.

layout
The vertex placement algorithm which is according to igraph package.

...
System reserved (no specific usage).

Author(s)

Reza Mohammadi and Ernst Wit

References

Mohammadi, A. and E. Wit (2015). Bayesian Structure Learning in Sparse Gaussian Graphical Models, Bayesian Analysis, 10(1):109-138

Mohammadi, A. and E. Wit (2015). BDgraph: An R Package for Bayesian Structure Learning in Graphical Models, arXiv preprint arXiv:1501.05108

Dobra, A. and A. Mohammadi (2017). Loglinear Model Selection and Human Mobility, arXiv preprint arXiv:1711.02623

Mohammadi, A. et al (2017). Bayesian modelling of Dupuytren disease by using Gaussian copula graphical models, Journal of the Royal Statistical Society: Series C

Mohammadi, A., Massam H., and G. Letac (2017). The Ratio of Normalizing Constants for Bayesian Graphical Gaussian Model Selection, arXiv preprint arXiv:1706.04416

See Also

bdgraph

Examples

## Not run: 
# Generating multivariate normal data from a 'random' graph
data.sim <- bdgraph.sim( n = 50, p = 6, size = 7, vis = TRUE )
   
bdgraph.obj <- bdgraph( data = data.sim )
   
plot( bdgraph.obj )
   
bdgraph.obj <- bdgraph( data = data.sim, save.all = TRUE )
   
plot( bdgraph.obj, number.g = 4 )
  
plot( bdgraph.obj, cut = 0.4 )

## End(Not run)

[Package BDgraph version 2.49 Index]