plot.graph {BDgraph}	R Documentation

Plot function for S3 class "graph"

Description

Visualizes structure of the graph.

Usage

## S3 method for class 'graph'
plot( x, cut = 0.5, mode = "undirected", diag = FALSE, main = NULL, 
                      vertex.color = "white", vertex.label.color = 'black', ... )

Arguments

x
An object of S3 class "graph", from function graph.sim.

cut
This option is for the case where input 'x' is the object of class "bdgraph" or "ssgraph". Threshold for including the links in the selected graph based on the estimated posterior probabilities of the links.

mode
Type of graph which is according to R package igraph.

diag
Logical which is according to R package igraph.

main
Graphical parameter (see plot).

vertex.color
The vertex color which is according to R package igraph.

vertex.label.color
The vertex label color which is according to R package igraph.

...
System reserved (no specific usage).

Author(s)

Reza Mohammadi a.mohammadi@uva.nl

References

Mohammadi, R. and Wit, E. C. (2019). BDgraph: An R Package for Bayesian Structure Learning in Graphical Models, Journal of Statistical Software, 89(3):1-30

Mohammadi, A. and Wit, E. C. (2015). Bayesian Structure Learning in Sparse Gaussian Graphical Models, Bayesian Analysis, 10(1):109-138

Letac, G., Massam, H. and Mohammadi, R. (2018). The Ratio of Normalizing Constants for Bayesian Graphical Gaussian Model Selection, arXiv preprint arXiv:1706.04416v2

Dobra, A. and Mohammadi, R. (2018). Loglinear Model Selection and Human Mobility, Annals of Applied Statistics, 12(2):815-845

Mohammadi, A. et al (2017). Bayesian modelling of Dupuytren disease by using Gaussian copula graphical models, Journal of the Royal Statistical Society: Series C, 66(3):629-645

Mohammadi, A. and Dobra, A. (2017). The R Package BDgraph for Bayesian Structure Learning in Graphical Models, ISBA Bulletin, 24(4):11-16

See Also

graph.sim, bdgraph.sim

Examples

# Generating a 'random' graph 
adj <- graph.sim( p = 10, graph = "random" )
plot( adj )
adj

[Package BDgraph version 2.62 Index]