plotcoda {BDgraph}	R Documentation

Convergence plot

Description

Visualizes the cumulative occupancy fractions of all possible links in the graph. It can be used for monitoring the convergence of the sampling algorithms, BDMCMC and RJMCMC.

Usage

 plotcoda( bdgraph.obj, thin = NULL, control = TRUE, main = NULL, ... ) 

Arguments

bdgraph.obj
An object of S3 class "bdgraph", from function bdgraph.

thin
An option for getting fast result for a cumulative plot according to part of the iteration.

control
Logical: if TRUE (default) and the number of nodes is greater than 15, then 100 links randomly is selected for visualization.

main
Graphical parameter (see plot).

...
System reserved (no specific usage).

Details

Note that a spending time for this function depends on the number of nodes. For fast result, you can choose bigger value for the 'thin' option.

Author(s)

Reza Mohammadi and Ernst Wit

References

Mohammadi, A. and E. Wit (2015). Bayesian Structure Learning in Sparse Gaussian Graphical Models, Bayesian Analysis, 10(1):109-138

Mohammadi, A. and E. Wit (2015). BDgraph: An R Package for Bayesian Structure Learning in Graphical Models, arXiv preprint arXiv:1501.05108

Dobra, A. and A. Mohammadi (2017). Loglinear Model Selection and Human Mobility, arXiv preprint arXiv:1711.02623

Mohammadi, A. et al (2017). Bayesian modelling of Dupuytren disease by using Gaussian copula graphical models, Journal of the Royal Statistical Society: Series C

Mohammadi, A., Massam H., and G. Letac (2017). The Ratio of Normalizing Constants for Bayesian Graphical Gaussian Model Selection, arXiv preprint arXiv:1706.04416

See Also

bdgraph

Examples

## Not run: 
# Generating multivariate normal data from a 'circle' graph
data.sim <- bdgraph.sim( n = 50, p = 6, graph = "circle", vis = TRUE )
  
bdgraph.obj <- bdgraph( data = data.sim, iter = 10000, burnin = 0 , save.all = TRUE )
   
plotcoda( bdgraph.obj )  

## End(Not run)

[Package BDgraph version 2.49 Index]