print.bdgraph {BDgraph}	R Documentation

Print function for S3 class "bdgraph"

Description

Prints the information about the selected graph which could be a graph with links for which their estimated posterior probabilities are greater than 0.5 or graph with the highest posterior probability. It provides adjacency matrix, size and posterior probability of the selected graph.

Usage

## S3 method for class 'bdgraph'
print( x, round = 2, ... )

Arguments

x
An object of S3 class "bdgraph", from function bdgraph.

round
A value to round the probabilities to the specified number of decimal places.

...
System reserved (no specific usage).

Author(s)

Reza Mohammadi and Ernst Wit

References

Mohammadi, A. and E. Wit (2015). Bayesian Structure Learning in Sparse Gaussian Graphical Models, Bayesian Analysis, 10(1):109-138

Mohammadi, A. and E. Wit (2015). BDgraph: An R Package for Bayesian Structure Learning in Graphical Models, arXiv preprint arXiv:1501.05108

Dobra, A. and A. Mohammadi (2017). Loglinear Model Selection and Human Mobility, arXiv preprint arXiv:1711.02623

Mohammadi, A. et al (2017). Bayesian modelling of Dupuytren disease by using Gaussian copula graphical models, Journal of the Royal Statistical Society: Series C

Mohammadi, A., Massam H., and G. Letac (2017). The Ratio of Normalizing Constants for Bayesian Graphical Gaussian Model Selection, arXiv preprint arXiv:1706.04416

See Also

bdgraph

Examples

## Not run: 
# Generating multivariate normal data from a 'random' graph
data.sim <- bdgraph.sim( n = 50, p = 6, vis = TRUE )
  
bdgraph.obj <- bdgraph( data = data.sim )
   
print( bdgraph.obj )

## End(Not run)

[Package BDgraph version 2.49 Index]