reinis {BDgraph}	R Documentation

Risk factors of coronary heart disease

Description

The dataset consist of 6 discrete variables as the potential risk factors of coronary heart disease. The data collected from 1841 men employed of a car factory in Czechoslovakia (Reinis et al. 1981).

Usage

 data( reinis ) 

Format

The format is a matrix with 1841 rows (number of individuals) and 6 column (number of variables).

References

Edwards and Havranek (1985). A fast procedure for model search in multidimensional contingency tables. Biometrika. 72:339-351

Reinis et al (1981). Prognostic significance of the risk profile in the prevention of coronary heart disease. Bratis. lek. Listy. 76:137-150

Mohammadi, A. and A. Dobra (2017). The R Package BDgraph for Bayesian Structure Learning in Graphical Models, ISBA Bulletin, 24(4):11-16

Examples

data( reinis )
dim( reinis )
head( reinis )

[Package BDgraph version 2.49 Index]