rmvnorm {BDgraph}	R Documentation

Generate data from the multivariate Normal distribution

Description

Random generation function from the multivariate Normal distribution with mean equal to mean and covariance matrix sigma.

Usage

 
    rmvnorm( n = 10, mean = rep( 0, length = ncol( sigma ) ), 
             sigma = diag( length( mean ) ) ) 

Arguments

n
Number of observations.

mean
Mean vector, default is rep( 0, length = ncol( sigma ) ).

sigma
positive definite covariance matrix, default is diag( length( mean ) ) .

Value

A numeric matrix with rows equal to n and columns equal to length( mean ).

Author(s)

Reza Mohammadi

See Also

bdgraph.sim, rwish, rgwish

Examples

    ## Not run: 
        mean  <- c( 5, 20 )       
        sigma <- matrix( c( 4, 2,
                            2, 5 ), 2, 2 )  # covariance matrix
        
        dat <- rmvnorm( n = 1000, mean = mean, sigma = sigma )
        plot( dat )
    
## End(Not run)

[Package BDgraph version 2.49 Index]