summary.bdgraph {BDgraph}	R Documentation

Summary function for S3 class "bdgraph"

Description

Provides a summary of the results for function bdgraph.

Usage

## S3 method for class 'bdgraph'
summary( object, round = 2, vis = TRUE, ... )

Arguments

object
An object of S3 class "bdgraph", from function bdgraph.

round
A value for rounding all probabilities to the specified number of decimal places.

vis
Visualize the results.

...
System reserved (no specific usage).

Value

selected_g
The adjacency matrix corresponding to the selected graph which has the highest posterior probability.

p_links
An upper triangular matrix corresponding to the posterior probabilities of all possible links.

K_hat
The estimated precision matrix.

Author(s)

Reza Mohammadi a.mohammadi@uva.nl and Ernst Wit

References

Mohammadi, R. and Wit, E. C. (2019). BDgraph: An R Package for Bayesian Structure Learning in Graphical Models, Journal of Statistical Software, 89(3):1-30

Mohammadi, A. and Wit, E. C. (2015). Bayesian Structure Learning in Sparse Gaussian Graphical Models, Bayesian Analysis, 10(1):109-138

Letac, G., Massam, H. and Mohammadi, R. (2018). The Ratio of Normalizing Constants for Bayesian Graphical Gaussian Model Selection, arXiv preprint arXiv:1706.04416v2

Dobra, A. and Mohammadi, R. (2018). Loglinear Model Selection and Human Mobility, Annals of Applied Statistics, 12(2):815-845

Mohammadi, A. et al (2017). Bayesian modelling of Dupuytren disease by using Gaussian copula graphical models, Journal of the Royal Statistical Society: Series C, 66(3):629-645

See Also

bdgraph, bdgraph.mpl

Examples

## Not run: 
# Generating multivariate normal data from a 'random' graph
data.sim <- bdgraph.sim( n = 50, p = 6, size = 7, vis = TRUE )
   
bdgraph.obj <- bdgraph( data = data.sim )
   
summary( bdgraph.obj )
   
bdgraph.obj <- bdgraph( data = data.sim, save = TRUE )
   
summary( bdgraph.obj )
   
summary( bdgraph.obj, vis = FALSE )

## End(Not run)

[Package BDgraph version 2.62 Index]