traceplot {BDgraph}	R Documentation

Trace plot of graph size

Description

Trace plot for graph size for the objects of S3 class "bdgraph", from function bdgraph. It is a tool for monitoring the convergence of the sampling algorithms, BDMCMC and RJMCMC.

Usage

 traceplot ( bdgraph.obj, acf = FALSE, pacf = FALSE, main = NULL, ... ) 

Arguments

bdgraph.obj
An object of S3 class "bdgraph", from function bdgraph. It also can be an object of S3 class "ssgraph", from the function ssgraph of R package ssgraph.

acf
Visualize the autocorrelation functions for graph size.

pacf
Visualize the partial autocorrelations for graph size.

main
Graphical parameter (see plot).

...
System reserved (no specific usage).

Author(s)

Reza Mohammadi a.mohammadi@uva.nl and Ernst Wit

References

Mohammadi, R. and Wit, E. C. (2019). BDgraph: An R Package for Bayesian Structure Learning in Graphical Models, Journal of Statistical Software, 89(3):1-30

Mohammadi, A. and Wit, E. C. (2015). Bayesian Structure Learning in Sparse Gaussian Graphical Models, Bayesian Analysis, 10(1):109-138

Dobra, A. and Mohammadi, R. (2018). Loglinear Model Selection and Human Mobility, Annals of Applied Statistics, 12(2):815-845

Mohammadi, A. et al (2017). Bayesian modelling of Dupuytren disease by using Gaussian copula graphical models, Journal of the Royal Statistical Society: Series C, 66(3):629-645

Mohammadi, A. and Dobra, A. (2017). The R Package BDgraph for Bayesian Structure Learning in Graphical Models, ISBA Bulletin, 24(4):11-16

See Also

plotcoda, bdgraph, bdgraph.mpl

Examples

## Not run: 
# Generating multivariate normal data from a 'random' graph
data.sim <- bdgraph.sim( n = 50, p = 6, size = 7, vis = TRUE )
   
bdgraph.obj <- bdgraph( data = data.sim, iter = 10000, burnin = 0, save = TRUE )
   
traceplot( bdgraph.obj )
   
traceplot( bdgraph.obj, acf = TRUE, pacf = TRUE )

## End(Not run)

[Package BDgraph version 2.62 Index]