transfer {BDgraph}	R Documentation

transfer for discrete data

Description

Transfers discrete data, by counting the duplicated rows.

Usage

 transfer( r_data ) 

Arguments

r_data
An (n x p) matrix or a data.frame corresponding to the data (n is the sample size and p is the number of variables).

Value

data
An (n \times p+1) matrix of transferred data, in which the last column is the frequency of duplicated rows.

Author(s)

Reza Mohammadi and Adrian Dobra

References

Mohammadi, A. and A. Dobra (2017). The R Package BDgraph for Bayesian Structure Learning in Graphical Models, ISBA Bulletin, 24(4):11-16

Dobra, A. and A. Mohammadi (2017). Loglinear Model Selection and Human Mobility, arXiv preprint arXiv:1711.02623

See Also

bdgraph.mpl and bdgraph.sim

Examples

## Not run: 
# Generating multivariate binary data from a 'random' graph
data.sim <- bdgraph.sim( n = 12, p = 4, size = 4, type = "discrete", cut = 2 )
r_data     <- data.sim $ data
r_data
   
# Transfer the data
transfer( r_data )
  

## End(Not run)

[Package BDgraph version 2.49 Index]