BDgraph-package {BDgraph}	R Documentation

Bayesian Structure Learning in Graphical Models

Description

The R package BDgraph provides statistical tools for Bayesian structure learning in undirected graphical models. The package is implemented the recent improvements in the Bayesian graphical models literature, including Mohammadi and Wit (2015), Mohammadi et al. (2017), Dobra and Mohammadi (2017). The computationally intensive tasks of the package are implemented in parallel using OpenMP in C++ and interfaced with R, to speed up the computations. Besides, the package contains several functions for simulation and visualization, as well as three multivariate datasets taken from the literature.

Details

The package includes following main functions:

bdgraph        Search algorithm in graphical models
bdgraph.mpl    Search algorithm in graphical models using marginal pseudo-likehlihood
bdgraph.sim    Graph data simulation 
graph.sim      Graph simulation 
bdgraph.npn    Nonparametric transfer 
compare        Graph structure comparison 
plinks         Estimated posterior link probabilities
plotcoda       Convergence plot
plotroc        ROC plot
rgwish         Sampling from G-Wishart distribution
rwish          Sampling from Wishart distribution
select         Graph selection
traceplot      Trace plot of graph size

How to cite this package

Whenever using this package, please cite as

Mohammadi R. and E. C. Wit (2017). BDgraph: Bayesian Structure Learning in 
Graphical Models using Birth-Death MCMC, R package version 3.47, 
https://CRAN.R-project.org/package=BDgraph

Author(s)

Reza Mohammadi <a.mohammadi@rug.nl> and Ernst Wit

References

Mohammadi, A. and E. Wit (2015). Bayesian Structure Learning in Sparse Gaussian Graphical Models, Bayesian Analysis, 10(1):109-138

Mohammadi, A. and E. Wit (2015). BDgraph: An R Package for Bayesian Structure Learning in Graphical Models, arXiv preprint arXiv:1501.05108

Mohammadi, A. et al (2017). Bayesian modelling of Dupuytren disease by using Gaussian copula graphical models, Journal of the Royal Statistical Society: Series C

Dobra, A. and A. Mohammadi (2017). Loglinear Model Selection and Human Mobility, arXiv preprint arXiv:1711.02623

Mohammadi, A., Massam H., and G. Letac (2017). The Ratio of Normalizing Constants for Bayesian Graphical Gaussian Model Selection, arXiv preprint arXiv:1706.04416

Mohammadi, A. and A. Dobra (2017). The R Package BDgraph for Bayesian Structure Learning in Graphical Models, ISBA Bulletin, 24(4):11-16

Lenkoski, A. (2013). A direct sampler for G-Wishart variates, Stat, 2:119-128

[Package BDgraph version 2.48 Index]