compare {BDgraph}	R Documentation

Graph structure comparison

Description

This function provides several measures to assess the performance of the graphical structure learning.

Usage

 
compare( target, est, est2 = NULL, est3 = NULL, est4 = NULL, main = NULL, 
         vis = FALSE ) 

Arguments

target
An adjacency matrix corresponding to the true graph structure in which a_{ij}=1 if there is a link between notes i and j, otherwise a_{ij}=0. It can be an object with S3 class "sim" from function bdgraph.sim. It can be an object with S3 class "graph" from function graph.sim.

est, est2, est3, est4
An adjacency matrix corresponding to an estimated graph. It can be an object with S3 class "bdgraph" from function bdgraph. It can be an object of S3 class "ssgraph", from the function ssgraph of R package ssgraph. It can be an object of S3 class "select", from the function huge.select of R package huge. Options est2, est3 and est4 are for comparing two or more different approaches.

main
A character vector giving the names for the result table.

vis
Visualize the true graph and estimated graph structures.

Value

True positive
The number of correctly estimated links.

True negative
The number of true non-existing links which is correctly estimated.

False positive
The number of links which they are not in the true graph, but are incorrectly estimated.

False negative
The number of links which they are in the true graph, but are not estimated.

F1-score
A weighted average of the "positive predictive" and "true positive rate". The F1-score value reaches its best value at 1 and worst score at 0.

Specificity
The Specificity value reaches its best value at 1 and worst score at 0.

Sensitivity
The Sensitivity value reaches its best value at 1 and worst score at 0.

MCC
The Matthews Correlation Coefficients (MCC) value reaches its best value at 1 and worst score at 0.

Author(s)

Reza Mohammadi a.mohammadi@uva.nl, Antonio Abbruzzo, and Ivan Vujacic

References

Mohammadi, R. and Wit, E. C. (2019). BDgraph: An R Package for Bayesian Structure Learning in Graphical Models, Journal of Statistical Software, 89(3):1-30

Mohammadi, A. and Wit, E. C. (2015). Bayesian Structure Learning in Sparse Gaussian Graphical Models, Bayesian Analysis, 10(1):109-138

Mohammadi, A. et al (2017). Bayesian modelling of Dupuytren disease by using Gaussian copula graphical models, Journal of the Royal Statistical Society: Series C, 66(3):629-645

Letac, G., Massam, H. and Mohammadi, R. (2018). The Ratio of Normalizing Constants for Bayesian Graphical Gaussian Model Selection, arXiv preprint arXiv:1706.04416v2

Dobra, A. and Mohammadi, R. (2018). Loglinear Model Selection and Human Mobility, Annals of Applied Statistics, 12(2):815-845

See Also

bdgraph, bdgraph.mpl, bdgraph.sim, plotroc

Examples

## Not run: 
# Generating multivariate normal data from a 'random' graph
data.sim <- bdgraph.sim( n = 50, p = 6, size = 7, vis = TRUE )
    
# Running sampling algorithm based on GGMs 
sample.ggm <- bdgraph( data = data.sim, method = "ggm", iter = 10000 )
   
# Comparing the results
compare( data.sim, sample.ggm, main = c( "True", "GGM" ), vis = TRUE )
      
# Running sampling algorithm based on GCGMs
sample.gcgm <- bdgraph( data = data.sim, method = "gcgm", iter = 10000 )

# Comparing GGM and GCGM methods
compare( data.sim, sample.ggm, sample.gcgm, main = c( "True", "GGM", "GCGM" ), vis = TRUE )

## End(Not run)

[Package BDgraph version 2.62 Index]