pgraph {BDgraph}	R Documentation

Posterior probabilities of the graphs

Description

Provides the estimated posterior probabilities for the most likely graphs or a specific graph.

Usage

 pgraph( bdgraph.obj, number.g = 4, adj_g = NULL ) 

Arguments

bdgraph.obj
An object of S3 class "bdgraph", from function bdgraph.

number.g
The number of graphs with the highest posterior probabilities to be shown. This option is ignored if 'adj_g' is specified.

adj_g
An adjacency matrix corresponding to a graph structure. It is an upper triangular matrix in which a_{ij}=1 if there is a link between notes i and j, otherwise a_{ij}=0. It also can be an object of S3 class "sim", from function bdgraph.sim.

Value

selected_g
The graphs with the highest posterior probabilities.

prob_g
A vector of the posterior probabilities of the graphs corresponding to 'selected\_g'.

Author(s)

Reza Mohammadi and Ernst Wit

References

Mohammadi, A. and E. Wit (2015). Bayesian Structure Learning in Sparse Gaussian Graphical Models, Bayesian Analysis, 10(1):109-138

Mohammadi, A. and E. Wit (2015). BDgraph: An R Package for Bayesian Structure Learning in Graphical Models, arXiv preprint arXiv:1501.05108

Dobra, A. and A. Mohammadi (2017). Loglinear Model Selection and Human Mobility, arXiv preprint arXiv:1711.02623

Mohammadi, A. et al (2017). Bayesian modelling of Dupuytren disease by using Gaussian copula graphical models, Journal of the Royal Statistical Society: Series C

Mohammadi, A., Massam H., and G. Letac (2017). The Ratio of Normalizing Constants for Bayesian Graphical Gaussian Model Selection, arXiv preprint arXiv:1706.04416

See Also

bdgraph

Examples

## Not run: 
# Generating multivariate normal data from a 'random' graph
data.sim <- bdgraph.sim( n = 50, p = 6, size = 6, vis = TRUE )
   
bdgraph.obj <- bdgraph( data = data.sim, save.all = TRUE )
   
# Estimated posterior probability of the true graph
pgraph( bdgraph.obj, adj_g = data.sim )
   
# Estimated posterior probability of first and second graphs with highest probabilities
pgraph( bdgraph.obj, number.g = 2 )

## End(Not run)

[Package BDgraph version 2.49 Index]