plot.graph {BDgraph}	R Documentation

Plot function for S3 class "graph"

Description

Visualizes structure of the graph.

Usage

## S3 method for class 'graph'
plot( x, main = NULL, layout = layout.circle, ... )

Arguments

x
An object of S3 class "graph", from function graph.sim.

main
Graphical parameter (see plot).

layout
The vertex placement algorithm which is according to igraph package.

...
System reserved (no specific usage).

Author(s)

Reza Mohammadi and Ernst Wit

References

Mohammadi, A. and E. Wit (2015). Bayesian Structure Learning in Sparse Gaussian Graphical Models, Bayesian Analysis, 10(1):109-138

Mohammadi, A. and E. Wit (2015). BDgraph: An R Package for Bayesian Structure Learning in Graphical Models, arXiv preprint arXiv:1501.05108v2

Dobra, A. and A. Mohammadi (2017). Loglinear Model Selection and Human Mobility, arXiv preprint arXiv:1711.02623

Mohammadi, A. et al (2017). Bayesian modelling of Dupuytren disease by using Gaussian copula graphical models, Journal of the Royal Statistical Society: Series C

Mohammadi, A., Massam H., and G. Letac (2017). The Ratio of Normalizing Constants for Bayesian Graphical Gaussian Model Selection, arXiv preprint arXiv:1706.04416

Mohammadi, A. and A. Dobra (2017). The R Package BDgraph for Bayesian Structure Learning in Graphical Models, ISBA Bulletin, 24(4):11-16

See Also

graph.sim, bdgraph.sim

Examples

## Not run: 
# Generating a 'random' graph 
adj <- graph.sim( p = 10, graph = "random" )
plot( adj )
adj

## End(Not run)

[Package BDgraph version 2.49 Index]