plotroc {BDgraph}	R Documentation

ROC plot

Description

Draws the ROC curve according to the true graph structure for object of S3 class "bdgraph", from function bdgraph.

Usage

 
plotroc( sim.obj, bdgraph.obj, bdgraph.obj2 = NULL, bdgraph.obj3 = NULL, 
         bdgraph.obj4 = NULL, cut = 20, smooth = FALSE, label = TRUE, 
         main = "ROC Curve" ) 

Arguments

sim.obj
An object of S3 class "sim", from function bdgraph.sim. It also can be the adjacency matrix corresponding to the true graph structure in which a_{ij}=1 if there is a link between notes i and j, otherwise a_{ij}=0.

bdgraph.obj
An object of S3 class "bdgraph", from function bdgraph. It also can be an upper triangular matrix corresponding to the estimated posterior probabilities for all possible links.

bdgraph.obj2
An object of S3 class "bdgraph", from function bdgraph. It also can be an upper triangular matrix corresponding to the estimated posterior probabilities for all possible links. It is for comparing two different approaches.

bdgraph.obj3
An object of S3 class "bdgraph", from function bdgraph. It also can be an upper triangular matrix corresponding to the estimated posterior probabilities for all possible links. It is for comparing three different approaches.

bdgraph.obj4
An object of S3 class "bdgraph", from function bdgraph. It also can be an upper triangular matrix corresponding to the estimated posterior probabilities for all possible links. It is for comparing four different approaches.

cut
Number of cut points.

smooth
Logical: for smoothing the ROC curve.

label
Logical: for adding legend to the ROC plot.

main
An overall title for the plot.

Author(s)

Reza Mohammadi and Ernst Wit

References

Mohammadi, A. and E. Wit (2015). Bayesian Structure Learning in Sparse Gaussian Graphical Models, Bayesian Analysis, 10(1):109-138

Mohammadi, A. and E. Wit (2015). BDgraph: An R Package for Bayesian Structure Learning in Graphical Models, arXiv preprint arXiv:1501.05108

Dobra, A. and A. Mohammadi (2017). Loglinear Model Selection and Human Mobility, arXiv preprint arXiv:1711.02623

Mohammadi, A. et al (2017). Bayesian modelling of Dupuytren disease by using Gaussian copula graphical models, Journal of the Royal Statistical Society: Series C

Mohammadi, A., Massam H., and G. Letac (2017). The Ratio of Normalizing Constants for Bayesian Graphical Gaussian Model Selection, arXiv preprint arXiv:1706.04416

See Also

bdgraph and compare

Examples

## Not run: 
# Generating multivariate normal data from a 'random' graph
data.sim <- bdgraph.sim( n = 30, p = 6, size = 7, vis = TRUE )
   
# Runing sampling algorithm
bdgraph.obj <- bdgraph( data = data.sim, iter = 10000 )
# Comparing the results
plotroc( data.sim, bdgraph.obj )
   
# To compare the results based on CGGMs approach
bdgraph.obj2 <- bdgraph( data = data.sim, method = "gcgm", iter = 10000 )
# Comparing the resultss
plotroc( data.sim, bdgraph.obj, bdgraph.obj2, label = FALSE )
legend( "bottomright", c( "GGMs", "GCGMs" ), lty = c( 1,2 ), col = c( "black", "red" ) )   

## End(Not run)

[Package BDgraph version 2.48 Index]