summary.bdgraph {BDgraph}	R Documentation

Summary function for S3 class "bdgraph"

Description

Provides a summary of the results for function bdgraph.

Usage

## S3 method for class 'bdgraph'
summary( object, round = 2, vis = TRUE, ... )

Arguments

object
An object of S3 class "bdgraph", from function bdgraph.

round
A value for rounding all probabilities to the specified number of decimal places.

vis
Visualize the results.

...
System reserved (no specific usage).

Value

best.graph
The adjacency matrix corresponding to the selected graph which has the highest posterior probability.

p_links
An upper triangular matrix corresponding to the posterior probabilities of all possible links.

K_hat
The estimated precision matrix.

Author(s)

Reza Mohammadi and Ernst Wit

References

Mohammadi, A. and E. Wit (2015). Bayesian Structure Learning in Sparse Gaussian Graphical Models, Bayesian Analysis, 10(1):109-138

Mohammadi, A. and E. Wit (2015). BDgraph: An R Package for Bayesian Structure Learning in Graphical Models, arXiv preprint arXiv:1501.05108

Dobra, A. and A. Mohammadi (2017). Loglinear Model Selection and Human Mobility, arXiv preprint arXiv:1711.02623

Mohammadi, A. et al (2017). Bayesian modelling of Dupuytren disease by using Gaussian copula graphical models, Journal of the Royal Statistical Society: Series C

Mohammadi, A., Massam H., and G. Letac (2017). The Ratio of Normalizing Constants for Bayesian Graphical Gaussian Model Selection, arXiv preprint arXiv:1706.04416

See Also

bdgraph

Examples

## Not run: 
# Generating multivariate normal data from a 'random' graph
data.sim <- bdgraph.sim( n = 50, p = 6, size = 7, vis = TRUE )
   
bdgraph.obj <- bdgraph( data = data.sim )
   
summary( bdgraph.obj )
   
bdgraph.obj <- bdgraph( data = data.sim, save.all = TRUE )
   
summary( bdgraph.obj )
   
summary( bdgraph.obj, vis = FALSE )

## End(Not run)

[Package BDgraph version 2.49 Index]